third_party/blink/web_tests/virtual/http-cache-default-backend/external/wpt/fetch/http-cache/vary.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] HTTP cache doesn't invalidate existing Vary response
  assert_less_than: Response 3 does not come from cache expected a number less than 3 but got 3
Harness: the test ran to completion.
